+++ resotocore/tests/resotocore/task/task_handler_test.py
@@ -25,7 +25,7 @@ from resotocore.task.task_description import (
     Job,
     TaskSurpassBehaviour,
 )
-from core.task import TaskHandler
+from resotocore.task import TaskHandler
 from resotocore.task.task_handler import TaskHandlerService
 from tests.resotocore.db.entitydb import InMemoryDb
 
@@ -133,7 +133,7 @@ async def test_parse_job_line_event_and_time_trigger(task_handler: TaskHandler) 
     job = await task_handler.parse_job_line("test", '0 5 * * sat cleanup_plan : match t2 == "node" | clean')
     assert job.trigger == TimeTrigger("0 5 * * sat")
     assert job.command.command == 'match t2 == "node" | clean'
-    assert job.wait[0] == EventTrigger("cleanup_plan")  # type: ignore
+    assert job.wait[0] == EventTrigger("cleanup_plan")
     with pytest.raises(ParseError):
         await task_handler.parse_job_line("test", '0 5 invalid_cron_expr * sat cleanup_plan:match t2 == "node" | clean')
